refactor: extract MarkerWithLabel class
